organize plugins by functionality
