+++ src/plugins/parser/reactPartial.js
