+++ src/plugins/parser/resourcelink.js
