+++ src/plugins/parser/vocablink.js
