+++ src/plugins/process/renderRedactions.js
